Files are serializing, deserialize to json equally, load in RL and even play. But they crash RL after a bit, and the network data bytes have a ton of differences still

+++ RocketLeagueReplayParser/NetworkStream/UniqueId.cs
@@ -62,7 +62,7 @@ namespace RocketLeagueReplayParser.NetworkStream
 
         public void Serialize(BitWriter bw)
         {
-            bw.Write((UInt32)Type);
+            bw.Write((byte)Type);
             SerializeId(bw);
         }
